Initial spawner test
some interesting results, need to maybe work on the random nature. i am
using a very small subset of data. so will probably improve with more
objects etc. unit seems to be colliding with little edges that will have
to be fixed. but im happy with this for now work on it more tonight